os/drivers/wireless/scsc: Modify slsi api header directory
- Applying header location and refactoring
- Clean up scsc driver and slsi api code
