+++ apps/examples/testcase/ta_tc/device_management/itc/itc_dm_main.c
@@ -26,7 +26,7 @@
 #include <dm/dm_error.h>
 #include <dm/dm_connectivity.h>
 #include <protocols/dhcpc.h>
-#include <slsi_wifi/slsi_wifi_api.h>
+#include <tinyara/wifi/slsi/slsi_wifi_api.h>
 
 #include "tc_common.h"
 #include "itc_internal.h"
