+++ external/iotivity/iotivity_1.2-rel/resource/csdk/connectivity/src/ip_adapter/tizenrt/caipnwmonitor.c
@@ -33,7 +33,6 @@
 #include <net/if.h>
 #include <netdb.h>
 #include <errno.h>
-//#include <slsi_wifi/slsi_wifi_api.h>
 
 #include "octhread.h"
 #include "caipnwmonitor.h"
