+++ framework/src/dm/conn/wifi/slsi_dm_connectivity.c
@@ -41,8 +41,7 @@
 #include <dm/dm_connectivity.h>
 
 #include <protocols/dhcpc.h>
-#include <slsi_wifi/slsi_wifi_api.h>
-#include <slsi_wifi/slsi_wifi_utils.h>
+#include <tinyara/wifi/slsi/slsi_wifi_api.h>
 #include <arpa/inet.h>
 
 #define DM_GET_INTERFACE_INFO 0
