+++ os/drivers/wireless/scsc/Kconfig
@@ -6,6 +6,51 @@ menuconfig SCSC_WLAN
 	select PIPES
 
 if SCSC_WLAN
+config SLSI_WIFI_DEFAULT_WLAN_COUNTRY_CODE
+	string "Country code (2 chars)"
+	default "00"
+	---help---
+		The country code to use for the board when starting the first time.
+		The country code is a two char upper-case value, e.g. DK, KR, JP, US.
+		Use "00" for multi-country international.
+
+config SLSI_WIFI_DEFAULT_WLAN_TX_POWER
+	int "TX power in dBm (12-30)"
+	range 12 30
+	default 30
+	---help---
+		The default tx power value in dBm to use for the board when starting the first time.
+		The tx power value is in the range 0-30 and is of the type dBm.
+
+config SLSI_WIFI_FILESYSTEM_SUPPORT
+	bool "Support filesystem"
+	default n
+	---help---
+		The API can utilize the filesystem for configuration files and log files if needed.
+		This is the default behavior but can be disabled on systems without filesystems.
+		If filesystem support is disabled then functionality like auto-reconnect will not work as
+		this uses the config file to store the relevant info across reboots.
+
+if SLSI_WIFI_FILESYSTEM_SUPPORT
+config SLSI_WIFI_DIR
+	string "Path to where wifi can store the config files (ommit the trailing /)"
+	default "/mnt/wifi"
+	---help---
+		The path where the wifi system can store the config and log files if needed. The path must be
+		without the trailing /.
+
+config SLSI_WIFI_LOG_FILE_NAME
+	string "Name of the log file to use"
+	default "slsiwifi.log"
+	---help---
+		The name of the file used to store logs from the wifi system.
+
+config SLSI_WIFI_CONFIG_FILE_NAME
+	string "name of the configuration file to use"
+	default "slsiwifi.conf"
+	---help---
+		The name of the file used to store information about the wifi system.
+endif
 
 config SLSI_RX_PERFORMANCE_TEST
 	bool "Enable UDP RX performance test"
@@ -16,7 +61,7 @@ config SLSI_RX_PERFORMANCE_TEST
 
 config SCSC_TX_FLOW_CONTROL
 	bool "Enable TX flow control"
-        default n
+        default y
 	---help---
 	  This option enables TX flow control
 
@@ -70,14 +115,14 @@ endif
 
 menuconfig SCSC_CORE
 	bool "SCSC Core driver Module support"
-	default n
+	default y
 	select SPI
 
 if SCSC_CORE
 
 config SCSC_PLATFORM
 	bool "Samsung SCSC MIF Platform driver implementation"
-	default n
+	default y
 
 config SCSC_WLANLITE
 	bool "Samsung SCSC WLANLITE driver mode"
