+++ os/drivers/wireless/scsc/Make.defs
@@ -18,6 +18,9 @@
 
 ifeq ($(CONFIG_SCSC_WLAN),y)
 
+#Driver API
+CSRCS += slsi_wifi_api.c
+
 #Upper driver
 CSRCS += \
 	dev.c \
@@ -62,11 +65,15 @@ CSRCS += \
 
 ifeq ($(CONFIG_LWNL80211_SLSI),y)
 CSRCS += slsi_drv.c
+else
+CSRCS += slsi_utils.c
 endif
 
 # Include wireless devices build support
 
 DEPPATH += --dep-path wireless$(DELIM)scsc
 VPATH += :wireless$(DELIM)scsc
+CFLAGS += ${shell $(INCDIR) $(INCDIROPT) "$(CC)" $(TOPDIR)$(DELIM)..$(DELIM)external$(DELIM)wpa_supplicant$(DELIM)src}
 CFLAGS += ${shell $(INCDIR) $(INCDIROPT) "$(CC)" $(TOPDIR)$(DELIM)drivers$(DELIM)wireless$(DELIM)scsc}
+
 endif
