+++ os/include/tinyara/wifi/slsi/slsi_wifi_api.h
@@ -31,6 +31,12 @@
 #ifndef SLSI_WIFI_API_H_
 #define SLSI_WIFI_API_H_
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdint.h>
+#include <unistd.h>
+#include <tinyara/mm/mm.h>
+
 #ifdef __cplusplus
 #define EXTERN extern "C"
 extern "C"
@@ -46,11 +52,23 @@ extern "C"
 #define CTRL_IFNAME "wl1"
 #endif
 
-
-
 #ifndef BIT
 #define BIT(x) (1 << (x))
 #endif
+
+/* Supported security modes as string */
+#define SLSI_WIFI_SECURITY_OPEN          "open"
+#define SLSI_WIFI_SECURITY_WEP_OPEN      "wep"
+#define SLSI_WIFI_SECURITY_WEP_SHARED    "wep_shared"
+#define SLSI_WIFI_SECURITY_WPA_MIXED     "wpa_mixed"
+#define SLSI_WIFI_SECURITY_WPA_TKIP      "wpa_tkip"
+#define SLSI_WIFI_SECURITY_WPA_AES       "wpa_aes"
+#define SLSI_WIFI_SECURITY_WPA2_MIXED    "wpa2_mixed"
+#define SLSI_WIFI_SECURITY_WPA2_TKIP     "wpa2_tkip"
+#define SLSI_WIFI_SECURITY_WPA2_AES      "wpa2_aes"
+
+#define MODE_STRING_MAX 100
+#define MAX_SSID_LEN    (4 * 32 + 1)
 
 typedef enum {
 	SLSI_SEC_MODE_OPEN			= 0,				// 00000000
@@ -429,6 +447,15 @@ int8_t WiFiSaveConfig(void);
  *   Will set external autoconnect enabled or disabled
  */
 int8_t WiFiSetAutoconnect(uint8_t check);
+
+/*return a security config struct depending on the mode, and the input security string (see above
+ * defines.
+ */
+slsi_security_config_t *getSecurityConfig(char *sec_type, char *psk, WiFi_InterFace_ID_t mode);
+/* prints a nicely ordered list of the scan results from the scan_info list
+ */
+
+void printScanResult(slsi_scan_info_t *list);
 
 #undef EXTERN
 #ifdef  __cplusplus
